compiling server actions and resources
