+++ examples/with-auth/src/routes/login.tsx
@@ -1,11 +1,9 @@
-import { createAction, FormError } from "solid-start/form";
-import server, { redirect } from "solid-start/server";
+import { redirect, createServerResource, createServerAction } from "solid-start/server";
 import { db } from "~/db";
 import { createUserSession, getUser, login, register } from "~/db/session";
-import { useParams, useRouteData } from "solid-app-router";
 import { Show } from "solid-js";
-import { ErrorBoundary } from "solid-start/error-boundary";
-import { createRouteResource } from "solid-start/router";
+import ErrorBoundary from "solid-start/error-boundary";
+import { useParams, useRouteData, FormError } from "solid-start/router";
 
 function validateUsername(username: unknown) {
   if (typeof username !== "string" || username.length < 3) {
@@ -20,14 +18,12 @@ function validatePassword(password: unknown) {
 }
 
 export function routeData() {
-  return createRouteResource(
-    server(async function (ctx) {
-      if (await getUser(ctx.request)) {
-        throw redirect("/");
-      }
-      return {};
-    })
-  );
+  return createServerResource(async ({ request }) => {
+    if (await getUser(request)) {
+      throw redirect("/");
+    }
+    return {};
+  });
 }
 
 export default function Login() {
@@ -38,61 +34,59 @@ export default function Login() {
    * statements, while still returning the accurate HTTP status, 400 Bad Request,
    * to the client.
    */
-  const loginAction = createAction(
-    server(async (form: FormData) => {
-      const loginType = form.get("loginType");
-      const username = form.get("username");
-      const password = form.get("password");
-      const redirectTo = form.get("redirectTo") || "/";
-      if (
-        typeof loginType !== "string" ||
-        typeof username !== "string" ||
-        typeof password !== "string" ||
-        typeof redirectTo !== "string"
-      ) {
-        throw new FormError(`Form not submitted correctly.`);
-      }
+  const loginAction = createServerAction(async (form: FormData) => {
+    const loginType = form.get("loginType");
+    const username = form.get("username");
+    const password = form.get("password");
+    const redirectTo = form.get("redirectTo") || "/";
+    if (
+      typeof loginType !== "string" ||
+      typeof username !== "string" ||
+      typeof password !== "string" ||
+      typeof redirectTo !== "string"
+    ) {
+      throw new FormError(`Form not submitted correctly.`);
+    }
 
-      const fields = { loginType, username, password };
-      const fieldErrors = {
-        username: validateUsername(username),
-        password: validatePassword(password)
-      };
-      if (Object.values(fieldErrors).some(Boolean)) {
-        throw new FormError("Fields invalid", { fieldErrors, fields });
-      }
+    const fields = { loginType, username, password };
+    const fieldErrors = {
+      username: validateUsername(username),
+      password: validatePassword(password)
+    };
+    if (Object.values(fieldErrors).some(Boolean)) {
+      throw new FormError("Fields invalid", { fieldErrors, fields });
+    }
 
-      switch (loginType) {
-        case "login": {
-          const user = await login({ username, password });
-          if (!user) {
-            throw new FormError(`Username/Password combination is incorrect`, {
-              fields
-            });
-          }
-          return createUserSession(`${user.id}`, redirectTo);
+    switch (loginType) {
+      case "login": {
+        const user = await login({ username, password });
+        if (!user) {
+          throw new FormError(`Username/Password combination is incorrect`, {
+            fields
+          });
         }
-        case "register": {
-          const userExists = await db.user.findUnique({ where: { username } });
-          if (userExists) {
-            throw new FormError(`User with username ${username} already exists`, {
-              fields
-            });
-          }
-          const user = await register({ username, password });
-          if (!user) {
-            throw new FormError(`Something went wrong trying to create a new user.`, {
-              fields
-            });
-          }
-          return createUserSession(`${user.id}`, redirectTo);
-        }
-        default: {
-          throw new FormError(`Login type invalid`, { fields });
-        }
+        return createUserSession(`${user.id}`, redirectTo);
       }
-    })
-  );
+      case "register": {
+        const userExists = await db.user.findUnique({ where: { username } });
+        if (userExists) {
+          throw new FormError(`User with username ${username} already exists`, {
+            fields
+          });
+        }
+        const user = await register({ username, password });
+        if (!user) {
+          throw new FormError(`Something went wrong trying to create a new user.`, {
+            fields
+          });
+        }
+        return createUserSession(`${user.id}`, redirectTo);
+      }
+      default: {
+        throw new FormError(`Login type invalid`, { fields });
+      }
+    }
+  });
 
   const params = useParams();
   return (
@@ -141,7 +135,7 @@ export default function Login() {
               </div>
               <Show when={loginAction.submission("login")?.error}>
                 <p class="text-red-400" role="alert" id="error-message">
-                  {loginAction.submissions()["login"]?.error.message}
+                  {loginAction.submission("login")?.error.message}
                 </p>
               </Show>
               <button
