+++ fixtures/cloudflare-workers/src/routes/action/index.tsx
@@ -1,31 +1,24 @@
-import server, { redirect } from "solid-start/server";
-import { createAction, createForm } from "solid-start/form";
-import { createComputed, createResource } from "solid-js";
+import server, { redirect, createServerResource, createServerAction } from "solid-start/server";
+import { createComputed } from "solid-js";
 import { getUser, logout } from "~/db/session";
-import { useNavigate, useRouteData } from "solid-app-router";
+import { useNavigate, useRouteData } from "solid-start/router";
 
 export function routeData() {
-  return createResource(
-    server(async () => {
-      if (!(await getUser(server.request))) {
-        throw redirect("/action/login");
-      }
+  return createServerResource(async () => {
+    if (!(await getUser(server.request))) {
+      throw redirect("/action/login");
+    }
 
-      return {};
-    })
-  );
+    return {};
+  });
 }
 
 export default function Home() {
-  const [data] = useRouteData<ReturnType<typeof routeData>>();
+  const data = useRouteData<ReturnType<typeof routeData>>();
   createComputed(data);
   const navigate = useNavigate();
 
-  const [_, logoutAction] = createAction(
-    server(async () => {
-      return await logout(server.request);
-    })
-  );
+  const [_, logoutAction] = createServerAction(() => logout(server.request));
 
   return (
     <main class="w-full p-4 space-y-2">
