+++ packages/start/router/Form.tsx
@@ -1,7 +1,7 @@
 // Credits to the Remix team for the Form implementation:
 // https://github.com/remix-run/remix/blob/main/packages/remix-react/components.tsx#L865
 
-import { ActionSubmission } from "./createAction";
+import { ActionSubmission } from "./createRouteAction";
 import { ComponentProps, createEffect, mergeProps, onCleanup, splitProps, JSX } from "solid-js";
 
 export interface FormAction<Data> {
