+++ packages/start/server/resource.ts
@@ -1,3 +1,4 @@
-import { createRouteResource } from "../router";
+import { createRouteResource, createRouteAction } from "../router";
 
 export const createServerResource = createRouteResource;
+export const createServerAction = createRouteAction;
